requirements.txt
streamlit
numpy
faiss-cpu
requests
sentence-transformers
rank_bm25
watchdog
nltk